refactor remove scenes-api from slobs 0.16.3
