+++ app/components/windows/RenameSource.vue.ts
@@ -3,7 +3,7 @@ import { Component } from 'vue-property-decorator';
 import { Inject } from '../../services/core/injector';
 import ModalLayout from '../ModalLayout.vue';
 import { WindowsService } from '../../services/windows';
-import { IScenesServiceApi } from '../../services/scenes';
+import { ScenesService } from '../../services/scenes';
 import { ISourcesServiceApi } from '../../services/sources';
 import { $t } from 'services/i18n';
 
@@ -12,7 +12,7 @@ import { $t } from 'services/i18n';
 })
 export default class RenameSource extends Vue {
   @Inject() sourcesService: ISourcesServiceApi;
-  @Inject() scenesService: IScenesServiceApi;
+  @Inject() scenesService: ScenesService;
   @Inject() windowsService: WindowsService;
 
   options: {
