+++ app/services/scenes/scene-node.ts
@@ -2,7 +2,7 @@
  * abstract class for representing scene's folders and items
  */
 import { ServiceHelper, mutation } from '../core/stateful-service';
-import { TSceneNodeType } from './scenes-api';
+import { TSceneNodeType } from './scenes';
 import { Inject } from '../core/injector';
 import { ScenesService, Scene, ISceneItemNode, SceneItemFolder, SceneItem, TSceneNode } from './index';
 import { SelectionService } from 'services/selection';
