+++ app/services/scenes/scenes.ts
@@ -1,28 +1,105 @@
 import Vue from 'vue';
 import { without } from 'lodash';
-import { StatefulService, mutation } from 'services/core/stateful-service';
+import { Subject } from 'rxjs';
+import { mutation, StatefulService } from 'services/core/stateful-service';
 import { TransitionsService } from 'services/transitions';
 import { WindowsService } from 'services/windows';
-import {
-  IScene,
-  Scene,
-  ISceneItem,
-  SceneItem,
-  IScenesState,
-  ISceneCreateOptions,
-  IScenesServiceApi
-} from './index';
-import { SourcesService, ISource } from 'services/sources';
-import electron from 'electron';
-import { Subject } from 'rxjs';
+import { Scene, SceneItem } from './index';
+import { ISource, SourcesService, ISourceAddOptions } from 'services/sources';
 import { Inject } from 'services/core/injector';
 import * as obs from '../../../obs-api';
 import { $t } from 'services/i18n';
 import namingHelpers from 'util/NamingHelpers';
 import uuid from 'uuid/v4';
 
-export class ScenesService extends StatefulService<IScenesState> implements IScenesServiceApi {
+export type TSceneNodeModel = ISceneItem | ISceneItemFolder;
 
+export interface IScene extends IResource {
+  id: string;
+  name: string;
+  nodes: (ISceneItem | ISceneItemFolder)[];
+}
+
+export interface ISceneNodeAddOptions {
+  id?: string; // A new ID will be assigned if one is not provided
+  sourceAddOptions?: ISourceAddOptions;
+  select?: boolean; // Immediately select this source
+  initialTransform?: IPartialTransform;
+}
+
+export interface ISceneItemInfo {
+  id: string;
+  sourceId: string;
+  x: number;
+  y: number;
+  scaleX: number;
+  scaleY: number;
+  visible: boolean;
+  crop: ICrop;
+  locked?: boolean;
+  rotation?: number;
+}
+
+export interface IScenesState {
+  activeSceneId: string;
+  displayOrder: string[];
+  scenes: Dictionary<IScene>;
+}
+
+export interface ISceneCreateOptions {
+  duplicateSourcesFromScene?: string;
+  sceneId?: string; // A new ID will be generated if one is not provided
+  makeActive?: boolean;
+}
+
+export interface ITransform {
+  position: IVec2;
+  scale: IVec2;
+  crop: ICrop;
+  rotation: number;
+}
+
+export interface IPartialTransform {
+  position?: Partial<IVec2>;
+  scale?: Partial<IVec2>;
+  crop?: Partial<ICrop>;
+  rotation?: number;
+}
+
+export interface ISceneItemSettings {
+  transform: ITransform;
+  visible: boolean;
+  locked: boolean;
+}
+
+export interface IPartialSettings {
+  transform?: IPartialTransform;
+  visible?: boolean;
+  locked?: boolean;
+}
+
+export interface ISceneItem extends ISceneItemSettings, ISceneItemNode {
+  sceneItemId: string;
+  sourceId: string;
+  obsSceneItemId: number;
+  sceneNodeType: 'item';
+}
+
+export type TSceneNodeType = 'item' | 'folder';
+
+export interface ISceneItemNode extends IResource {
+  id: string;
+  sceneId: string;
+  sceneNodeType: TSceneNodeType;
+  parentId?: string;
+}
+
+export interface ISceneItemFolder extends ISceneItemNode {
+  name: string;
+  sceneNodeType: 'folder';
+}
+
+export class ScenesService extends StatefulService<IScenesState> {
   static initialState: IScenesState = {
     activeSceneId: '',
     displayOrder: [],
